requirements.txt
tkcalendar>=1.6.1
playsound>=1.3.0
pyinstaller>=6.0.0 
# pygobject>=3.40.0 # Recommended for playsound on Linux, best installed via system package manager (e.g., apt install python3-gi)

# For Gemini AI SDK
google-generativeai

# For Speech-to-Text
# SpeechRecognition
# PyAudio (if needed by SpeechRecognition)

# For Text-to-Speech
pyttsx3 # (cross-platform) or gTTS (requires internet)

# For potential embedded web view (YouTube integration)
# cefpython3 (Windows/Linux) or tkhtmlview (simpler, limited)
# Consider platform-specific dependencies if chosen
